test: merge config/context enhanced tests into main test files, remove _enhanced_test.go
